pyproject.toml
[project]
name = "social"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "nltk>=3.9.2",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "vk-api>=11.10.0",
]
